Split pruning into two files
